Refactors AppChip to shared component
Moves the AppChip component to a shared directory for reusability.

Improves the user loading experience by adding a linear progress indicator.
